+++ R/gg_blanket.R
@@ -1150,19 +1150,36 @@ gg_blanket <- function(
             reverse = col_legend_rev,
             ncol = col_legend_ncol,
             nrow = col_legend_nrow
-          ),
-
-          shape = ggplot2::guide_legend(
-            reverse = col_legend_rev,
-            ncol = col_legend_ncol,
-            nrow = col_legend_nrow
-          ),
-          linetype = ggplot2::guide_legend(
-            reverse = col_legend_rev,
-            ncol = col_legend_ncol,
-            nrow = col_legend_nrow
           )
         )
+
+#ONLY GUIDE LEGEND IF AES THE SAME AS COLOUR
+        #   alpha = ggplot2::guide_legend(
+        #     reverse = col_legend_rev,
+        #     ncol = col_legend_ncol,
+        #     nrow = col_legend_nrow
+        #   ),
+        #   shape = ggplot2::guide_legend(
+        #     reverse = col_legend_rev,
+        #     ncol = col_legend_ncol,
+        #     nrow = col_legend_nrow
+        #   ),
+        #   size = ggplot2::guide_legend(
+        #     reverse = col_legend_rev,
+        #     ncol = col_legend_ncol,
+        #     nrow = col_legend_nrow
+        #   ),
+        #   linewidth = ggplot2::guide_legend(
+        #     reverse = col_legend_rev,
+        #     ncol = col_legend_ncol,
+        #     nrow = col_legend_nrow
+        #   ),
+        #   linetype = ggplot2::guide_legend(
+        #     reverse = col_legend_rev,
+        #     ncol = col_legend_ncol,
+        #     nrow = col_legend_nrow
+        #   )
+        # )
     }
 
     #expand limits if necessary
@@ -1566,51 +1583,136 @@ gg_blanket <- function(
   #############################################################################
   # titles_to_case
   #############################################################################
+
   if (rlang::is_null(x_title)) {
-    if (stringr::str_detect(stat_name, "sf")) {
-      x_title <- ""
+    if ((!rlang::quo_is_null(x))) {
+      if (!rlang::is_null(attr(dplyr::pull(data, {{ x }}), "label"))) {
+        x_title <- attr(dplyr::pull(data, {{ x }}), "label")
+      }
+      else {
+        if (stringr::str_detect(stat_name, "sf")) {
+          x_title <- ""
+        }
+        else if (!rlang::is_null(plot_build$plot$labels$x)) {
+          x_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$x[1]), titles_to_case)
+        }
+      }
     }
-    else if (!rlang::is_null(plot_build$plot$labels$x)) {
-      x_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$x[1]), titles_to_case)
+    else {
+      if (stringr::str_detect(stat_name, "sf")) {
+        x_title <- ""
+      }
+      else if (!rlang::is_null(plot_build$plot$labels$x)) {
+        x_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$x[1]), titles_to_case)
+      }
     }
   }
+
   if (rlang::is_null(y_title)) {
-    if (stringr::str_detect(stat_name, "sf")) {
-      y_title <- ""
+    if ((!rlang::quo_is_null(y))) {
+      if (!rlang::is_null(attr(dplyr::pull(data, {{ y }}), "label"))) {
+        y_title <- attr(dplyr::pull(data, {{ y }}), "label")
+      }
+      else {
+        if (stringr::str_detect(stat_name, "sf")) {
+          y_title <- ""
+        }
+        else if (!rlang::is_null(plot_build$plot$labels$y)) {
+          y_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$y[1]), titles_to_case)
+        }
+      }
     }
-    else if (!rlang::is_null(plot_build$plot$labels$y)) {
-      y_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$y[1]), titles_to_case)
+    else {
+      if (stringr::str_detect(stat_name, "sf")) {
+        y_title <- ""
+      }
+      else if (!rlang::is_null(plot_build$plot$labels$y)) {
+        y_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$y[1]), titles_to_case)
+      }
     }
   }
 
   if (rlang::is_null(col_title)) {
-    if (!rlang::is_null(plot_build$plot$labels$fill)) {
-      col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$fill[1]), titles_to_case)
+    if ((!rlang::quo_is_null(col))) {
+      if (!rlang::is_null(attr(dplyr::pull(data, {{ col }}), "label"))) {
+        col_title <- attr(dplyr::pull(data, {{ col }}), "label")
+      }
+      else {
+        if (!rlang::is_null(plot_build$plot$labels$fill)) {
+          col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$fill[1]), titles_to_case)
+        }
+        else if (!rlang::is_null(plot_build$plot$labels$colour)) {
+          col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$colour[1]), titles_to_case)
+        }
+      }
     }
-    else if (!rlang::is_null(plot_build$plot$labels$colour)) {
-      col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$colour[1]), titles_to_case)
+    else {
+      if (!rlang::is_null(plot_build$plot$labels$fill)) {
+        col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$fill[1]), titles_to_case)
+      }
+      else if (!rlang::is_null(plot_build$plot$labels$colour)) {
+        col_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$colour[1]), titles_to_case)
+      }
     }
   }
 
   if (!rlang::is_null(plot_build$plot$labels$alpha)) {
-    alpha_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$alpha[1]), titles_to_case)
+    if (!rlang::is_null(plot_build$plot$labels$colour[1])) {
+      if (rlang::as_name(plot_build$plot$labels$colour[1]) == rlang::as_name(plot_build$plot$labels$alpha[1])) {
+        alpha_title <- col_title
+      }
+      else alpha_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$alpha[1]), titles_to_case)
+    }
+    else if (!rlang::is_null(plot_build$plot$labels$fill[1])) {
+      if (rlang::as_name(plot_build$plot$labels$fill[1]) == rlang::as_name(plot_build$plot$labels$alpha[1])) {
+        alpha_title <- col_title
+      }
+      else alpha_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$alpha[1]), titles_to_case)
+    }
+    else {
+      alpha_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$alpha[1]), titles_to_case)
+    }
   } else alpha_title <- NULL
 
-  if (!rlang::is_null(plot_build$plot$labels$shape)) {
-    shape_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$shape[1]), titles_to_case)
-  } else shape_title <- NULL
-
-  if (!rlang::is_null(plot_build$plot$labels$linetype)) {
-    linetype_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$linetype[1]), titles_to_case)
-  } else linetype_title <- NULL
-
-  if (!rlang::is_null(plot_build$plot$labels$size)) {
-    size_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$size[1]), titles_to_case)
-  } else size_title <- NULL
-
-  if (!rlang::is_null(plot_build$plot$labels$linewidth)) {
-    linewidth_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$linewidth[1]), titles_to_case)
-  } else linewidth_title <- NULL
+  # if (!rlang::is_null(plot_build$plot$labels$shape)) {
+  #   if (rlang::as_name(plot_build$plot$labels$colour[1]) == rlang::as_name(plot_build$plot$labels$shape[1]) |
+  #       rlang::as_name(plot_build$plot$labels$fill[1]) == rlang::as_name(plot_build$plot$labels$shape[1])) {
+  #     shape_title <- col_title
+  #   }
+  #   else {
+  #     shape_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$shape[1]), titles_to_case)
+  #   }
+  # } else shape_title <- NULL
+  #
+  # if (!rlang::is_null(plot_build$plot$labels$size)) {
+  #   if (rlang::as_name(plot_build$plot$labels$colour[1]) == rlang::as_name(plot_build$plot$labels$size[1]) |
+  #       rlang::as_name(plot_build$plot$labels$fill[1]) == rlang::as_name(plot_build$plot$labels$size[1])) {
+  #     size_title <- col_title
+  #   }
+  #   else {
+  #     size_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$size[1]), titles_to_case)
+  #   }
+  # } else size_title <- NULL
+  #
+  # if (!rlang::is_null(plot_build$plot$labels$linewidth)) {
+  #   if (rlang::as_name(plot_build$plot$labels$colour[1]) == rlang::as_name(plot_build$plot$labels$linewidth[1]) |
+  #       rlang::as_name(plot_build$plot$labels$fill[1]) == rlang::as_name(plot_build$plot$labels$linewidth[1])) {
+  #     linewidth_title <- col_title
+  #   }
+  #   else {
+  #     linewidth_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$linewidth[1]), titles_to_case)
+  #   }
+  # } else linewidth_title <- NULL
+  #
+  # if (!rlang::is_null(plot_build$plot$labels$linetype)) {
+  #   if (rlang::as_name(plot_build$plot$labels$colour[1]) == rlang::as_name(plot_build$plot$labels$linetype[1]) |
+  #       rlang::as_name(plot_build$plot$labels$fill[1]) == rlang::as_name(plot_build$plot$labels$linetype[1])) {
+  #     linetype_title <- col_title
+  #   }
+  #   else {
+  #     linetype_title <- purrr::map_chr(rlang::as_name(plot_build$plot$labels$linetype[1]), titles_to_case)
+  #   }
+  # } else linetype_title <- NULL
 
   plot <- plot +
     ggplot2::labs(
@@ -1618,46 +1720,15 @@ gg_blanket <- function(
       subtitle = subtitle,
       caption = caption,
       x = x_title,
-      y = y_title
+      y = y_title,
+      colour = col_title,
+      fill = col_title,
+      alpha = alpha_title,
+      # shape = shape_title,
+      # size = size_title,
+      # linewidth = linewidth_title,
+      # linetype = linetype_title
     )
-
-  if (!rlang::is_null(col_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        colour = col_title,
-        fill = col_title
-      )
-  }
-  if (!rlang::is_null(alpha_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        alpha = alpha_title
-      )
-  }
-  if (!rlang::is_null(shape_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        shape = shape_title
-      )
-  }
-  if (!rlang::is_null(linetype_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        linetype = linetype_title
-      )
-  }
-  if (!rlang::is_null(size_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        size = size_title
-      )
-  }
-  if (!rlang::is_null(linewidth_title)) {
-    plot <- plot +
-      ggplot2::labs(
-        linewidth = linewidth_title
-      )
-  }
 
   ##############################################################################
   # auto panel.grid, line & ticks removal